add validation server-side that user can't like their own messages

+++ app.py
@@ -316,17 +316,19 @@ def add_liked_message(message_id):
 
     message = Message.query.get_or_404(message_id)
 
-    #add validation that msg is not by user
+    if message.user is not g.user:
+        if message in g.user.likes:
+            LikedMessage.query.filter(LikedMessage.message_id == message_id, LikedMessage.user_id == g.user.id).delete()
+            db.session.commit()
 
-    if message in g.user.likes:
-        LikedMessage.query.filter(LikedMessage.message_id == message_id, LikedMessage.user_id == g.user.id).delete()
-        db.session.commit()
-
+        else: 
+            g.user.likes.append(message)
+            db.session.commit()
     else: 
-        g.user.likes.append(message)
-        db.session.commit()
-
+        flash("You can't like your own messages!")
+    
     return redirect(request.referrer)
+
